Try modelling sinon stub instances as a real type.
Now the identify function of the sinonStubInstance type actually returns
true for objects that only have spy values.

It means that the output of the type has to account for the "false
positives" by being generic enough that the error message for eg.

expect(spy, 'to have a call satisfying', {spy: spy});

... still makes sense.